+++ server/hazard-service/src/main/java/edu/illinois/ncsa/incore/service/hazard/controllers/HurricaneController.java
@@ -3,12 +3,15 @@ package edu.illinois.ncsa.incore.service.hazard.controllers;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import edu.illinois.ncsa.incore.common.AllocationConstants;
 import edu.illinois.ncsa.incore.common.HazardConstants;
 import edu.illinois.ncsa.incore.common.auth.IAuthorizer;
 import edu.illinois.ncsa.incore.common.auth.Privileges;
+import edu.illinois.ncsa.incore.common.dao.IAllocationRepository;
 import edu.illinois.ncsa.incore.common.dao.ISpaceRepository;
 import edu.illinois.ncsa.incore.common.exceptions.IncoreHTTPException;
 import edu.illinois.ncsa.incore.common.models.Space;
+import edu.illinois.ncsa.incore.common.utils.AllocationUtils;
 import edu.illinois.ncsa.incore.common.utils.UserInfoUtils;
 import edu.illinois.ncsa.incore.service.hazard.dao.IHurricaneRepository;
 import edu.illinois.ncsa.incore.service.hazard.exception.UnsupportedHazardException;
@@ -57,6 +60,9 @@ public class HurricaneController {
 
     @Inject
     private ISpaceRepository spaceRepository;
+
+    @Inject
+    private IAllocationRepository allocationRepository;
 
     @Inject
     private IAuthorizer authorizer;
@@ -148,6 +154,18 @@ public class HurricaneController {
         @ApiParam(hidden = true) @FormDataParam("hurricane") String hurricaneJson,
         @ApiParam(hidden = true) @FormDataParam("file") List<FormDataBodyPart> fileParts) {
 
+        // check if the user's number of the hazard is within the allocation
+        if (!AllocationUtils.canCreateHazard(allocationRepository, spaceRepository, this.username)) {
+            throw new IncoreHTTPException(Response.Status.FORBIDDEN,
+                AllocationConstants.HAZARD_ALLOCATION_MESSAGE);
+        }
+
+        // check if the user's number of the hazard dataset is within the allocation
+        if (!AllocationUtils.canCreateHazardDataset(allocationRepository, spaceRepository, this.username)) {
+            throw new IncoreHTTPException(Response.Status.FORBIDDEN,
+                AllocationConstants.HAZARD_DATASET_ALLOCATION_MESSAGE);
+        }
+
         ObjectMapper mapper = new ObjectMapper();
         Hurricane hurricane = null;
         try {
@@ -198,6 +216,10 @@ public class HurricaneController {
                     spaceRepository.addSpace(space);
 
                     hurricane.setSpaces(spaceRepository.getSpaceNamesOfMember(hurricane.getId()));
+
+                    // add one more dataset in the usage
+                    AllocationUtils.increaseNumHazards(spaceRepository, this.username);
+
                     return hurricane;
                 } else {
                     throw new IncoreHTTPException(Response.Status.BAD_REQUEST, "Could not create hurricane, no files were attached with " +
@@ -326,6 +348,9 @@ public class HurricaneController {
                     spaceRepository.addSpace(space);
                 }
             }
+
+            // reduce the number of hazard from the space
+            AllocationUtils.reduceNumHazard(spaceRepository, this.username);
 
             return deletedHurr;
         } else {